+++ tests/pure/index.js
@@ -144,6 +144,7 @@ import './esnext.array.is-template-object';
 import './esnext.async-iterator.constructor';
 import './esnext.async-iterator.drop';
 import './esnext.async-iterator.filter';
+import './esnext.async-iterator.for-each';
 import './esnext.async-iterator.from';
 import './esnext.async-iterator.map';
 import './esnext.async-iterator.take';