+++ tests/tests/esnext.async-iterator.for-each.js
@@ -1,0 +1,25 @@
+import { createIterator } from '../helpers/helpers';
+
+QUnit.test('AsyncIterator#forEach', assert => {
+  assert.expect(10);
+  const async = assert.async();
+  const { forEach } = AsyncIterator.prototype;
+
+  assert.isFunction(forEach);
+  assert.arity(forEach, 1);
+  assert.name(forEach, 'forEach');
+  assert.looksNative(forEach);
+  assert.nonEnumerable(AsyncIterator.prototype, 'forEach');
+
+  const array = [];
+
+  forEach.call(createIterator([1, 2, 3]), it => array.push(it)).then(() => {
+    assert.arrayEqual(array, [1, 2, 3], 'basic functionality');
+    async();
+  });
+
+  assert.throws(() => forEach.call(undefined, () => { /* empty */ }), TypeError);
+  assert.throws(() => forEach.call(null, () => { /* empty */ }), TypeError);
+  assert.throws(() => forEach.call({}, () => { /* empty */ }), TypeError);
+  assert.throws(() => forEach.call([], () => { /* empty */ }), TypeError);
+});